Cleanup ItemStack serialization and deserialization
